+++ webapi/src/test/scala/org/knora/webapi/e2e/v1/ResourcesV1E2ESpec.scala
@@ -27,10 +27,9 @@ import akka.pattern._
 import akka.util.Timeout
 import org.knora.webapi.LiveActorMaker
 import org.knora.webapi.e2e.E2ESpec
-import org.knora.webapi.messages.v1respondermessages.resourcemessages.{CreateResourceApiRequestV1, CreateResourceValueV1}
-import org.knora.webapi.messages.v1respondermessages.triplestoremessages.RdfDataObject
-import org.knora.webapi.messages.v1respondermessages.valuemessages.CreateFileV1
-import org.knora.webapi.messages.v1storemessages.triplestoremessages.{RdfDataObject, ResetTriplestoreContent}
+import org.knora.webapi.messages.v1.responder.resourcemessages.{CreateResourceApiRequestV1, CreateResourceValueV1}
+import org.knora.webapi.messages.v1.responder.valuemessages.CreateFileV1
+import org.knora.webapi.messages.v1.store.triplestoremessages.{RdfDataObject, ResetTriplestoreContent}
 import org.knora.webapi.responders._
 import org.knora.webapi.responders.v1.ResponderManagerV1
 import org.knora.webapi.routing.v1.ResourcesRouteV1
