+++ webapi/src/test/scala/org/knora/webapi/responders/v1/HierarchicalListsResponderV1Spec.scala
@@ -25,10 +25,8 @@ import akka.actor.Props
 import akka.testkit._
 import org.knora.webapi._
 import org.knora.webapi.messages.v1.responder.listmessages._
-import org.knora.webapi.messages.v1respondermessages.listmessages._
-import org.knora.webapi.messages.v1respondermessages.triplestoremessages._
-import org.knora.webapi.messages.v1respondermessages.usermessages.{UserDataV1, UserProfileV1}
-import org.knora.webapi.messages.v1storemessages.triplestoremessages.{RdfDataObject, ResetTriplestoreContent, ResetTriplestoreContentACK}
+import org.knora.webapi.messages.v1.responder.usermessages.{UserDataV1, UserProfileV1}
+import org.knora.webapi.messages.v1.store.triplestoremessages.{RdfDataObject, ResetTriplestoreContent, ResetTriplestoreContentACK}
 import org.knora.webapi.store.{STORE_MANAGER_ACTOR_NAME, StoreManager}
 
 import scala.concurrent.duration._
