+++ webapi/src/test/scala/org/knora/webapi/responders/v1/ResourcesResponderV1SpecContextData.scala
@@ -22,8 +22,8 @@ package org.knora.webapi.responders.v1
 
 import akka.actor.ActorSystem
 import org.knora.webapi.Settings
-import org.knora.webapi.messages.v1respondermessages.resourcemessages._
-import org.knora.webapi.messages.v1respondermessages.usermessages.UserDataV1
+import org.knora.webapi.messages.v1.responder.resourcemessages._
+import org.knora.webapi.messages.v1.responder.usermessages.UserDataV1
 
 object ResourcesResponderV1SpecContextData {
