+++ webapi/src/test/scala/org/knora/webapi/responders/v1/ResourcesResponderV1SpecFullData.scala
@@ -21,11 +21,10 @@
 package org.knora.webapi.responders.v1
 
 import akka.actor.ActorSystem
+import org.knora.webapi.messages.v1.responder.resourcemessages._
+import org.knora.webapi.messages.v1.responder.usermessages.UserDataV1
 import org.knora.webapi.messages.v1.responder.valuemessages._
 import org.knora.webapi.{IRI, Settings}
-import org.knora.webapi.messages.v1respondermessages.resourcemessages._
-import org.knora.webapi.messages.v1respondermessages.usermessages.UserDataV1
-import org.knora.webapi.messages.v1respondermessages.valuemessages._
 
 object ResourcesResponderV1SpecFullData {
