+++ webapi/src/test/scala/org/knora/webapi/responders/v1/ResponderManagerV1Spec.scala
@@ -22,8 +22,8 @@ package org.knora.webapi.responders.v1
 
 import akka.actor.Props
 import akka.testkit.{ImplicitSender, TestActorRef, TestProbe}
-import org.knora.webapi.messages.v1respondermessages.resourcemessages.ResourceInfoGetRequestV1
-import org.knora.webapi.messages.v1respondermessages.usermessages.{UserDataV1, UserProfileV1}
+import org.knora.webapi.messages.v1.responder.resourcemessages.ResourceInfoGetRequestV1
+import org.knora.webapi.messages.v1.responder.usermessages.{UserDataV1, UserProfileV1}
 import org.knora.webapi.responders._
 import org.knora.webapi.routing.Authenticator
 import org.knora.webapi.{CoreSpec, TestProbeMaker}
