+++ webapi/src/test/scala/org/knora/webapi/responders/v1/SearchResponderV1Spec.scala
@@ -24,10 +24,8 @@ import akka.actor.Props
 import akka.testkit._
 import org.knora.webapi._
 import org.knora.webapi.messages.v1.responder.searchmessages._
-import org.knora.webapi.messages.v1respondermessages.searchmessages._
-import org.knora.webapi.messages.v1respondermessages.triplestoremessages.RdfDataObject
-import org.knora.webapi.messages.v1respondermessages.usermessages.{UserDataV1, UserProfileV1}
-import org.knora.webapi.messages.v1storemessages.triplestoremessages.{RdfDataObject, ResetTriplestoreContent, ResetTriplestoreContentACK}
+import org.knora.webapi.messages.v1.responder.usermessages.{UserDataV1, UserProfileV1}
+import org.knora.webapi.messages.v1.store.triplestoremessages.{RdfDataObject, ResetTriplestoreContent, ResetTriplestoreContentACK}
 import org.knora.webapi.responders._
 import org.knora.webapi.store._
 import org.knora.webapi.util.MessageUtil
