+++ webapi/src/test/scala/org/knora/webapi/responders/v1/UsersResponderV1Spec.scala
@@ -27,9 +27,8 @@ package org.knora.webapi.responders.v1
 import akka.testkit.{ImplicitSender, TestActorRef}
 import akka.util.Timeout
 import org.knora.webapi.CoreSpec
-import org.knora.webapi.messages.v1respondermessages.triplestoremessages._
-import org.knora.webapi.messages.v1respondermessages.usermessages.{UserDataV1, UserProfileByUsernameGetRequestV1, UserProfileV1}
-import org.knora.webapi.messages.v1storemessages.triplestoremessages.{SparqlSelectResponse, SparqlSelectResponseBody, SparqlSelectResponseHeader}
+import org.knora.webapi.messages.v1.responder.usermessages.{UserDataV1, UserProfileByUsernameGetRequestV1, UserProfileV1}
+import org.knora.webapi.messages.v1.store.triplestoremessages.{SparqlSelectResponse, SparqlSelectResponseBody, SparqlSelectResponseHeader}
 import org.mindrot.jbcrypt.BCrypt
 
 import scala.concurrent.duration._
