+++ webapi/src/test/scala/org/knora/webapi/store/triplestore/AllTriplestoreSpec.scala
@@ -24,8 +24,7 @@ import akka.actor.Props
 import akka.testkit.ImplicitSender
 import com.typesafe.config.ConfigFactory
 import org.knora.webapi.SettingsConstants._
-import org.knora.webapi.messages.v1respondermessages.triplestoremessages._
-import org.knora.webapi.messages.v1storemessages.triplestoremessages._
+import org.knora.webapi.messages.v1.store.triplestoremessages._
 import org.knora.webapi.store.{StoreManager, _}
 import org.knora.webapi.{CoreSpec, LiveActorMaker}
