+++ webapi/src/test/scala/org/knora/webapi/util/DateUtilSpec.scala
@@ -18,11 +18,9 @@
  * License along with Knora.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package unit.org.knora.rapier.util
+package org.knora.webapi.util
 
 import org.knora.webapi.messages.v1.responder.valuemessages.{DateValueV1, JulianDayCountValueV1, KnoraCalendarV1, KnoraPrecisionV1}
-import org.knora.webapi.messages.v1respondermessages.valuemessages.DateValueV1
-import org.knora.webapi.util.DateUtilV1
 import org.scalatest._
 
 /**
